requirements.txt
flask
slack_sdk
notion_client
python-dotenv
gunicorn
